AOC 2015 - Day 4 - Java
This is the first one I actually had a bit of a struggle with.
I have absolutely no knowledge of hashing algorithms and I have
very little information about bitwise operations/manipulation.
Everything I know, it's basically shown here. Funnily enough, 2nd
part was a lot easier than the 1st part because I could compare the
5th and 6th digits as opposed to 5th digit where I had to calculate
whether or not the 5th digit was a zero doing "bitwise AND".