AOC 2015 - Day 5 - Java
This one was just straight up annoying. It's not hard,
you just have to write a very specific regex pattern for
both of the solutions which I didn't (ChatGPT did). This
is not hard, just annoying (solely because of regex). So
I went and did it the long and needless way after the fact.