AOC 2015 - Day 6 - Java
This one was especially fun, but rather easy.
For the 2nd part, I tried to use a function
to determine the brightness change but due to
limitation of brightness never being negative,
this approach didn't work and I reverted to
using an array, but instead of storing whether
or not the light was on, I used it to determine
brightness using integers to represent it.