AOC 2015 - Day 7 - Java
While the past 6 puzzles have been generally
easy, with some being slightly hard due to me
not knowing about specific things (like hashing)
this one was seriously hard for me. I was stumped
and spent close to 8 hours, just thinking about it.
I knew I had to write a function that would recursively
find the values of everything but doing so was a pain
in the ass on multiple levels. At the end, the only
reason I could finish this is due to the fact that,
I went over instructions while debugging and found
the correct way to recurse over them.