AOC 2015 - Day 13 - Java
Adding myself at the end was very weird,
as I didn't account for anything like that
beforehand, I had to reformat the code to
be bearable. This is basically slightly more
complicated version of the 9th puzzle, but it's also
a lot more fun in my opinion. In both cases,
an algorithm with weights or something from
graph theory would make it much-much easier
probably.